% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Annotation.R
\name{Annotation}
\alias{Annotation}
\title{Annotation}
\description{
Provides an Annotation object.
This object will hold an annotation in a memory efficient manner.
Only the data that is needed will actually be loaded.
}
\section{Fields}{

\describe{
\item{\code{configurer}}{- holds global configuration data in Configurer object}

\item{\code{dao}}{- a reference to a DAO for expression data}
}}

