% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AnnotationDAO.R
\name{AnnotationDAO}
\alias{AnnotationDAO}
\title{AnnotationDAO}
\description{
Provides a generic annotation data access object.
Allows for a low memory footprint and efficient data access.
}
\section{Fields}{

\describe{
\item{\code{configurer}}{- holds global configuration data in Configurer object}

\item{\code{conn}}{- a connection to a temporary database holding the annotation}

\item{\code{table_name}}{- the name of the table to put data in}
}}

